Add RatingChange entity
Instead of CfRatingChange being stored in the repository, we separated it into two classes: CfRatingChange is the direct representation of what we get from the CF API, and RatingChange is what is actually stored (specifically, it also has the contest start time information, as well as anything else we might add in the future).

I might get rid of the contest repository, as it doesn't really seem needed anymore.